uploads section divided into api routes
