+++ src/app.js
@@ -5,7 +5,6 @@ const cors = require('cors');
 const path = require('path');
 
 const bodyParser = require('body-parser');
-const axios = require('axios');
 
 const fs = require('fs');
 
@@ -13,19 +12,12 @@ const app = express();
 const port = config.PORT;
 const host = config.HOST;
 
-//Essa variável determina o ambiente em que o sistema está rodanto // development | production
+//Essa variável determina o ambiente em que o sistema está rodando // development | production
 const environment = config.NODE_ENV;
-
-// Define o caminho absoluto para a pasta "public"
-const publicPath = path.join(__dirname, '../public');
 
 // Middleware para analisar o corpo das requisições JSON
 app.use(bodyParser.json({limit: '50mb'}));
 app.use(express.urlencoded({ extended: true, limit: '50mb' }));
-//app.use(express.json({verify: (req, res, buf) => { req.rawBody = buf.toString()},limit: '50mb'}));
-
-// Configura o middleware para servir arquivos estáticos
-app.use(express.static(publicPath));
 
 // Middleware para lidar com os cookies
 app.use(cookieParser());
@@ -39,71 +31,6 @@ app.use((req, res, next) => {
   next();
 });
 
-const multer = require("multer");
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    try {
-      let uploadPath;
-      if (req.body.type === 'banner') 
-        uploadPath = './public/uploads/posts_banner';
-      else if (req.body.type === 'photo')
-        uploadPath = './public/uploads/profile_photo';
-      else 
-        throw Error('Tipo de upload inválido (banner | photo)');
-      
-      // Verificar se o diretório existe, caso contrário, criá-lo
-      if (!fs.existsSync(uploadPath)) {
-        fs.mkdirSync(uploadPath, { recursive: true });
-      }
-
-      cb(null, uploadPath);
-
-    } catch (error) {
-      console.error('Erro :', error)
-    }
-  },
-  filename: function (req, file, cb) {
-    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9)
-    cb(null, file.fieldname + '-' + uniqueSuffix)
-  }
-});
-const upload = multer({ storage });
-
-// Upload com Multer
-app.post("/upload_files", upload.single("files"), uploadFiles);
-
-function uploadFiles(req, res) {
-    res.json({ message: "Successfully uploaded files", data: { 
-      'destination': req.file.destination,
-      'filename': req.file.filename 
-    }}
-  )};
-
-// Upload com axios
-app.post('/upload', async (req, res) => {
-  const { imageUrls, imageUris } = req.body;
-
-  // Iterar sobre as URLs das imagens e fazer o download e salvar localmente
-  for (let i = 0; i < imageUrls.length; i++) {
-    try {
-      const response = await axios.get(imageUrls[i], { responseType: 'arraybuffer' });
-      const directory = './public/uploads/posts_media';
-
-      // Verificar se o diretório existe, caso contrário, criá-lo
-      if (!fs.existsSync(directory)) {
-        fs.mkdirSync(directory, { recursive: true });
-      }
-
-      const filename = path.join(directory, imageUris[i]);
-      fs.writeFileSync(filename, Buffer.from(response.data, 'binary'));
-    } catch (error) {
-      console.error("Imagem não foi enviada!", error)
-    res.sendFile(path.join(publicPath, 'index.html'));
-  }}
-});
-
-
 // Middlewares que verificam se usuario está logado e seus derivados cargos
 const permissionVerify = require("./middlewares/permissionVerify.js")
 const adminPermissionVerify = require("./middlewares/adminPermissionVerify.js")
@@ -111,12 +38,14 @@ const editorPermissionVerify = require("./middlewares/editorPermissionVerify.js"
 
 //Base de rotas da API
 const routes = require('./routes');
-const { createToken } = require('./utils/tinymce-jwt.js');
 app.use('/api', routes);
 
 //Aqui é a parte que separa os ambientes para se trabalhar
 
 if(environment === "development"){
+
+  // Define o caminho absoluto para a pasta "public"
+  const publicPath = path.join(__dirname, '../public');
 
   //Servidor ssl para trabalhar localmente
   const https = require('https');
@@ -163,10 +92,6 @@ if(environment === "development"){
   app.get('/create', editorPermissionVerify, (req, res) => {
     res.sendFile(path.join(publicPath, 'index.html'));
   });
-
-  /* app.post('/tinymce-jwt', (req, res) => {
-  }); */
-  app.post('/tinymce-jwt', createToken);
 
   //Aqui é onde chama todas as outras rotas que não precisam de verificação especial
   app.get('*', permissionVerify, (req, res) => {
