+++ src/routes/index.js
@@ -7,6 +7,7 @@ const likesRoutes = require("./likesRoutes");
 const commentsRoutes = require("./commentsRoutes");
 const repliesRoutes = require("./repliesRoutes");
 const favoritePostsRoutes = require("./favoritePostsRoutes");
+const uploadsRoutes = require("./uploadsRoutes");
 
 router.use("/login", loginRoutes);
 router.use("/posts", postsRoutes);
@@ -14,5 +15,7 @@ router.use("/likes", likesRoutes);
 router.use("/comments", commentsRoutes);
 router.use("/replies", repliesRoutes);
 router.use("/favorite", favoritePostsRoutes);
+
+router.use("/uploads", uploadsRoutes);
 
 module.exports = router;
